+++ packages/json-schema-to-zod/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @composio/json-schema-to-zod
 
+## 0.1.9
+
+### Patch Changes
+
+- Add host name support in SDK
+
+## 0.1.8
+
+### Patch Changes
+
+- 1ab34ef: Fix json schema support in tools
+
+## 0.1.7
+
+### Patch Changes
+
+- c8e89d5: Fix telemetry transport
+
+## 0.1.6
+
+### Patch Changes
+
+- 37a1f01: Feat better connected account creation flow
+
+## 0.1.5
+
+### Patch Changes
+
+- df31cc2: Fix json schema parsing
+
+## 0.1.2
+
+### Patch Changes
+
+- f943ba4: Export all the types from the core SDK
+
 ## 0.1.2
 
 ### Patch Changes